Improve detokenisation
A space is added before expanding a token following an identifier.

A space is added after a token with the "Conditional" flag which is
followed by an alphanumeric, "_", or another token.

Tokens in REM statements are no longer expanded, as this breaks
peeking at binary data stashed in REM statements.

Tokens in DATA statements are no longer expanded, as they don't get
retokenised when the program is fed back to BASIC.

Non-expanded tokens which aren't printable Unicode characters are
now mapped to an equivalent printable Unicode character.  Fixes #23.
Also related to #42.