Fix handling of assembler comments (#78)
The BBC User Guide documents `\` as introducing a comment in the BBC
BASIC assembler.

Owlet previously incorrectly also recognised `;` as introducing a
comment.  I think there's a misunderstanding here because the assembler
seems to mostly ignore anything it isn't expecting after an instruction
until a `:` or the end of the line.  As a result of this, it may appear
that `;` is a comment character since you can write:

ROR A ; Anagram of ROAR

but you can equally write:

ROR A | Anagram of ROAR

or:

ROR A Anagram of ROAR

The acid test here is whether a comment works in place of an
instruction, e.g.

[\ comment

In this situation only `\` works, because only `\` really introduces
a comment.