Update for Mastodon
Disable the tweet button for now.  The bot is no longer accepting
programs on twitter.  Needs hooking up for mastodon ideally, but
that's more complicated due to the federated aspect, and may not
be possible at all.  If not we can probably provide a "copy toot to
clipboard" or something like that.

Update links to the Bot and Dom's profiles.  It seems Matt doesn't
have one yet so I've left that be.

Increase the length limit to 500 characters to match Mastodon
(though it's not quite that simple as you need a #bbcmicrobot
hashtag too; that sort of existed before as you needed @bbcmicrobot
except you could avoid that by posting as a reply to the bot).

Drop the base2048 encoding bits as the bot no longer supports
that.

Inline tokenisation details rather than linking to an old rheolism
tweet in case twitter dies completely or starts purging old messages to
reduce costs.