Highlight invalid operators from other languages
To help users new to BBC BASIC we now highlight common operators
from other languages which aren't operators in BBC BASIC as invalid
and suggest replacements for them.

Previously `==`, `!=`, `<<` and `>>` were actually incorrectly in the
list of valid BBC BASIC operators (the last two are valid in e.g.
BASIC V on RISC OS, but not in BASIC II).

Fixes #32