Drop invalid blockComment setting
There isn't a block comment in BBC BASIC, but we were setting
`REM` ... `:` as one.  The result was that the editor's "Toggle block
comment" command would insert a REM before the selection and ":" after
which doesn't actually comment out the selected block.

If we don't set this, the block comment commands just don't do
anything, which seems more helpful than doing something wrong.