Normalise smartquotes on import
BBCMicrobot normalises double smartquotes to ASCII double quote
before running the program.  Owlet had partial support for this
already in that syntax highlighting treated smartquotes as ASCII
but they weren't normalised before trying to run the program so
failed when RUN.

We now normalise on import so the program in the editor no longer
contains smartquotes and the syntax highlighting rules are simpler.

We don't seem to have existing tests for importing a BASIC program
but this URL provides a manual test:

https://bbcmic.ro/#{%22v%22%3A1%2C%22program%22%3A%22P.%E2%80%9Chi%E2%80%9D%22}