Quick quickfix fixes (#130)
* Update quickfix for Monaco >= 0.34.0

Quickfix was not doing anything in a production build, and failing
with "unsupported edit" in a debug build.

The problem was the `edit` attribute was renamed to `textEdit` in
Monaco 0.34.0.

* Fix quickfixes for invalid operators

We need to look up the replacement operator in this case rather than
just trying to upper case it (which does nothing for these operators).